World view makes sense for non-logged in users now.
Removed the displaying of the share box and the timeline and replies
tabs from the world view when a user is not logged in. Also removed from
the dashboard as this view is used when people are viewing hashtags and
it makes sense for non-users to be able to view hashtag content.