Fixed Twitter signup.
Changed test to reflect removal of /users/confirm

All services used to create accounts will go through /users/new with
varying messages.

The Authorizations are not being set in User, marked that for
further investigation.

Also added annotations for new tests that should be written to
keep an eye on twitter and facebook logins (not just signups)