Fixed remote subscription routes.
There are no tests for this because of how damned hard they are to
write. What parts of the specification do you mock? etc. (especially
with the asynchronous communication for subscribing to a hub)

Will try to mock the expected responses sometime soon so we have tests
for this route.