Adds check for profile url in webfinger user xrd.
Another node or ostatus client will want a link to the profile of the
user on our system. This is given within a profile-page webfinger field.
This test checks to see if that matches the user's profile page on the
site.

This can also be used for discovery, but that should be tested
separately from webfinger.