Adds salmon links to xrd.
These endpoints are completed in the routes, but without them in the
xrd, no ostatus client would know how to reach an rstat.us node user.
That would make me sad. All three links are required, but normally they
all point to the same route. This is true for us and status.net, so
don't be alarmed by it. I'm not sure why anybody would want to
distinguish them by route since the envelope exposes this, and this
distinction does not make it a separate resource at all. :(