Adds tests to ensure that the salmon urls exist in the user xrd.
I added a section to the webfinger tests which test that the xrd served
by a webfinger request contain the three salmon urls necessary to get
salmon notifications. Again, the three are not necessary, but it is nice
to provide all three even if they are the same url.

I will add more tests for the remaining things in the xrd to this set of
acceptance tests.