Adds use_ssl option to Author class which denotes whether or not the given Author wants https to be used by third-parties.
Now, we use this to generate urls seen by the remote nodes in the xrd
and generating urls for various models. The Author defaults, when
created for a local Author, to the protocol used to create it. If https
is used upon creation, https is used to pull feeds, etc.

This is logical since you generally care about these things when you
create a new account.