Adds use_ssl test and ensures that it is set by default using the domain field.
It will be set using the domain as long as the domain leaks the
protocol. It is kinda weird how we normalize it... but if you specify
https, it will set the use_ssl to true. otherwise, it sets it to false.
If you create the author using an https post, then it will use https
urls for the author.