Restricts exceptions handled to just Redfinger.
So, I agree with @carols10cents about this change. We could have a thing where an Author is rechecked and thus place the salmon notification in the delayed_job queue. That's a different issue. This is just about getting salmon notifications and webfinger working under https.

I'm ok with it not doing the queuing just for denial of service reasons... but we could create an issue for it?