Makes email duplication validation independent of username.
Oh, and it also ignores a nil value. You have no idea how frustrating it
was to see it take two nil values and rub them together and say "A horse
is a horse" and throw them in my face.

I also note that it was storing "" as an email sometimes, adding to the
fun time frustration hour.