Adds rels for user profile links to followers, following, and updates.
According to ALPS, the following rels are used on <a> tags:

messages-me: Applied to an A tag. A reference to a list representation
of all the messages posted by the designated user.

users-friends: Applied to an A tag. A reference to list representation
of the designated user's friend users.

users-followers: Applied to an A tag. A reference to list representation
of the users who follow the designated user.