Adds a multiuser/singleuser switch and admin users.
The first user to sign-up gets the admin bit. It is currently a secret
page "/admin". It is secret only through the fact that there is no link
to it. The multiuser switch will not allow any new accounts.

There is no way to re-assign admin or add new admins at the moment.

Look at that though, that meaningless helper somebody added
'admin_only!' is now actually meaningful. :)